+++ packages/ui/src/views/w3m-connecting-view/index.ts
@@ -1,26 +1,18 @@
 import { CoreUtil } from '@web3modal/core'
 import { LitElement, html } from 'lit'
 import { customElement } from 'lit/decorators.js'
-import { DataUtil } from '../../utils/DataUtil'
 import { ThemeUtil } from '../../utils/ThemeUtil'
+import { UiUtil } from '../../utils/UiUtil'
+import styles from './styles.css'
 
 @customElement('w3m-connecting-view')
 export class W3mConnectingView extends LitElement {
-  public static styles = [ThemeUtil.globalCss]
+  public static styles = [ThemeUtil.globalCss, styles]
 
   // -- private ------------------------------------------------------ //
-  private getSupportedPlatfroms() {
-    const { id, desktop, mobile } = CoreUtil.getConnectingRouterData()
-    const injectedWallets = DataUtil.injectedWallets()
-    const isInjected = injectedWallets.some(wallet => wallet.id === id)
-    const isDesktop = Boolean(desktop?.native ?? desktop?.universal)
-    const isMobile = Boolean(mobile?.native ?? mobile?.universal)
-
-    return { isInjected, isDesktop, isMobile }
-  }
 
   private desktopTemplate() {
-    const { isInjected, isDesktop, isMobile } = this.getSupportedPlatfroms()
+    const { isInjected, isDesktop, isMobile } = UiUtil.getCachedRouterWalletPlatforms()
 
     if (isInjected) {
       return html`<w3m-injected-connecting></w3m-injected-connecting>`
@@ -36,7 +28,7 @@ export class W3mConnectingView extends LitElement {
   }
 
   private mobileTemplate() {
-    const { isMobile } = this.getSupportedPlatfroms()
+    const { isMobile } = UiUtil.getCachedRouterWalletPlatforms()
 
     if (isMobile) {
       return html`<w3m-mobile-connecting></w3m-mobile-connecting>`
