+++ packages/ui/src/views/w3m-connecting-view/styles.css
@@ -1,0 +1,5 @@
+w3m-info-footer {
+  display: flex;
+  width: 100%;
+  justify-content: space-around;
+}
